# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/contact": {
      "post": {
        "operationId": "GmailController_sendMess",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Gmail"
        ]
      }
    },
    "/users/register": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/login": {
      "post": {
        "operationId": "UsersController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/admin/{id}": {
      "post": {
        "operationId": "UsersController_DoUserAdmin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/GetBalance/{userid}": {
      "get": {
        "operationId": "UsersController_getBalance",
        "parameters": [
          {
            "name": "userid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/PostBalance": {
      "post": {
        "operationId": "UsersController_setBalance",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/postBalanceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/roles/GetAllRoles": {
      "get": {
        "operationId": "RolesController_GetAllRoles",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ]
      }
    },
    "/favourites/createFavourite": {
      "post": {
        "operationId": "FavouritesController_CreateFavourite",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFavouriteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Favourites"
        ]
      }
    },
    "/favourites/getFavouritesFromUser/{userId}": {
      "get": {
        "operationId": "FavouritesController_GetFavouritesFromUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Favourites"
        ]
      }
    },
    "/favourites/deleteFavourite": {
      "delete": {
        "operationId": "FavouritesController_DeleteFavourite",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFavouriteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Favourites"
        ]
      }
    },
    "/favourites/getAllFavourites": {
      "get": {
        "operationId": "FavouritesController_GetAllFavouritesFromUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Favourites"
        ]
      }
    },
    "/products/GetProdById/{product_Id}": {
      "get": {
        "operationId": "ProductsController_getProductById",
        "parameters": [
          {
            "name": "product_Id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/products/GetAllProducts": {
      "get": {
        "operationId": "ProductsController_getAllProducts",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/products/SetProduct": {
      "post": {
        "operationId": "ProductsController_createProduct",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/products/DeleteProduct/{product_Id}": {
      "delete": {
        "operationId": "ProductsController_deleteProduct",
        "parameters": [
          {
            "name": "product_Id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/orders/getOrderById/{Id}": {
      "get": {
        "operationId": "OrdersController_getOrderById",
        "parameters": [
          {
            "name": "Id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/getAllOrders": {
      "get": {
        "operationId": "OrdersController_getAllOrders",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/createOrder": {
      "post": {
        "operationId": "OrdersController_createOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/deleteOrder": {
      "delete": {
        "operationId": "OrdersController_deleteOrder",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/rates/createRate": {
      "post": {
        "operationId": "RatesController_createRate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostRateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Rates"
        ]
      }
    },
    "/rates/deleteRate/{rating_id}": {
      "delete": {
        "operationId": "RatesController_deleteRate",
        "parameters": [
          {
            "name": "rating_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Rates"
        ]
      }
    },
    "/sort": {
      "get": {
        "operationId": "SortController_getSort",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Sort"
        ]
      }
    },
    "/rating/sendRating": {
      "post": {
        "operationId": "SyteRatingController_SendMess",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "SyteRating"
        ]
      }
    },
    "/rating/getRating": {
      "get": {
        "operationId": "SyteRatingController_GetRat",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "SyteRating"
        ]
      }
    }
  },
  "info": {
    "title": "API Documentation",
    "description": "API description",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "api",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "SendMessDto": {
        "type": "object",
        "properties": {}
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {}
      },
      "ValidateUserDto": {
        "type": "object",
        "properties": {}
      },
      "postBalanceDto": {
        "type": "object",
        "properties": {}
      },
      "CreateFavouriteDto": {
        "type": "object",
        "properties": {}
      },
      "SendProductDto": {
        "type": "object",
        "properties": {}
      },
      "SendOrderDto": {
        "type": "object",
        "properties": {}
      },
      "PostRateDto": {
        "type": "object",
        "properties": {}
      },
      "SendMessageDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}